Update submit unit tests
